# internal/adapters/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package repo

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Order struct {
	ID         int64              `json:"id"`
	CustomerID int64              `json:"customer_id"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type OrderItem struct {
	ID         int64              `json:"id"`
	OrderID    int64              `json:"order_id"`
	ProductID  int64              `json:"product_id"`
	Quantity   int32              `json:"quantity"`
	PriceCents int64              `json:"price_cents"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type Product struct {
	ID             int64              `json:"id"`
	Name           string             `json:"name"`
	PriceInCenters int32              `json:"price_in_centers"`
	Quantity       int32              `json:"quantity"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
}