# openapi.yaml
openapi: 3.0.3
info:
  title: E-commerce API
  description: API for managing e-commerce products and operations
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: health
    description: Health check endpoints
  - name: products
    description: Product management endpoints
  - name: orders
    description: Order management endpoints
paths:
  /:
    get:
      tags:
        - health
      summary: Root endpoint
      description: Returns a simple greeting message from the server
      operationId: getRoot
      responses:
        '200':
          description: Successful response
          content:
            text/plain:
              schema:
                type: string
                example: hi from server
  /health:
    get:
      tags:
        - health
      summary: Health check
      description: Returns the health status of the API
      operationId: healthCheck
      responses:
        '200':
          description: API is healthy
          content:
            text/plain:
              schema:
                type: string
                example: root.
  /api/v1/products:
    get:
      tags:
        - products
      summary: List all products
      description: Retrieves a list of all products in the catalog
      operationId: listProducts
      parameters:
        - name: limit
          in: query
          description: Maximum number of products to return
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of products to skip
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /product:
    post:
      tags:
        - products
      summary: Create a new product
      description: Creates a new product in the catalog
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /product/{id}:
    get:
      tags:
        - products
      summary: Get product by ID
      description: Retrieves a single product by its ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          description: Product ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orders/{id}:
    get:
      tags:
        - orders
      summary: Get order by ID
      description: Retrieves an order with aggregated data including items and calculated total price
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          description: Order ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the product
          example: 1
        name:
          type: string
          description: Name of the product
          example: Laptop
        priceInCenters:
          type: integer
          format: int64
          description: Price in cents
          example: 99999
        quantity:
          type: integer
          description: Available stock quantity
          example: 50
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the product was created
          example: '2025-01-15T10:30:00Z'
      required:
        - id
        - name
        - priceInCenters
        - quantity
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
          description: Name of the product
          example: Laptop
        priceInCents:
          type: integer
          format: int64
          description: Price in cents
          example: 99999
        quantity:
          type: integer
          description: Initial stock quantity
          example: 50
      required:
        - name
        - priceInCents
        - quantity
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the order
          example: 1
        customerId:
          type: integer
          format: int64
          description: Customer ID who placed the order
          example: 123
        totalPriceCents:
          type: integer
          format: int64
          description: Total price of the order in cents
          example: 199998
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
          description: List of items in the order
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the order was created
          example: '2025-01-15T10:30:00Z'
      required:
        - id
        - customerId
        - totalPriceCents
        - items
        - createdAt
    OrderItemResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          description: Product ID
          example: 1
        productName:
          type: string
          description: Name of the product
          example: Laptop
        quantity:
          type: integer
          description: Quantity ordered
          example: 2
        priceCents:
          type: integer
          format: int64
          description: Price per unit in cents
          example: 99999
        totalCents:
          type: integer
          format: int64
          description: Total price for this item in cents
          example: 199998
      required:
        - productId
        - productName
        - quantity
        - priceCents
        - totalCents
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Internal server error
        message:
          type: string
          description: Detailed error description
          example: Failed to retrieve products from database
      required:
        - error
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
security: []